角色的树状图，菜单的树状图展示以及上级菜单的获取。

+++ interface/demo-manager-interface/src/main/java/com/cmpay/yx/dto/MenuDTO.java
@@ -23,7 +23,7 @@ public class MenuDTO {
     /**
      * @Fields superMenuId 上级菜单ID 默认0
      */
-    private Integer superMenuId;
+    private Long superMenuId;
     /**
      * @Fields menuName 菜单名称
      */
@@ -64,4 +64,5 @@ public class MenuDTO {
      * @Fields child 子菜单
      */
     private List<MenuDTO> children;
+
 }
